add mock s3 server for tests
