+++ packages/repo/test/helpers/bootstrap.ts
@@ -3,6 +3,7 @@ import axios, { AxiosInstance } from 'axios';
 import { repoServer } from './repo-server';
 import { ipfsServerMock } from './ipfs-server-mock';
 import { redisServerMock } from './redis-server-mock';
+import { s3ServerMock } from './s3-server-mock';
 
 export function bootstrap() {
   const ctx = {} as {
@@ -11,13 +12,13 @@ export function bootstrap() {
     ipfsMockGet: (cid: string) => Buffer | undefined;
     ipfsMockRemove: (cid: string) => Promise<void>;
     ipfsMockClear: () => void;
+    s3List: () => Promise<string[]>;
   };
 
   before(async function () {
-    const [{ ipfsUrl, ipfsMockAdd, ipfsMockGet, ipfsMockRemove, ipfsMockClear }, { redisUrl }] = await Promise.all([
-      ipfsServerMock(),
-      redisServerMock(),
-    ]);
+    const { ipfsUrl, ipfsMockAdd, ipfsMockGet, ipfsMockRemove, ipfsMockClear } = await ipfsServerMock();
+    const { redisUrl } = await redisServerMock();
+    const { s3Url, s3List } = await s3ServerMock();
 
     const { repoUrl } = await repoServer({ redisUrl, ipfsUrl });
 
@@ -31,6 +32,7 @@ export function bootstrap() {
     ctx.ipfsMockGet = ipfsMockGet;
     ctx.ipfsMockRemove = ipfsMockRemove;
     ctx.ipfsMockClear = ipfsMockClear;
+    ctx.s3List = s3List;
   });
 
   return ctx;
